refactor: Separate backend classroom API
- move `joinClassroom`, `leaveClassroom` and `getClassrooms` from `chatController.js` to `classroomController.js`.
- integrate classroom controller functions into matching route endpoints in `classroomRoutes.js`.
- integrate `classroomRoutes` into the matching classroom API endpoint in `router.js`.
