+++ backend/routes/chat/chatRoutes.js
@@ -7,9 +7,6 @@ import {
   sendClassroomMessage,
   getUserMessages,
   getClassroomMessages,
-  getClassrooms,
-  joinClassroom,
-  leaveClassroom,
 } from "../../controller/chat/chatController.js";
 
 chatRoutes
@@ -19,14 +16,6 @@ chatRoutes
 
 // chatRoutes.route("/classroom/:id").post(protect, sendClassroomMessage);
 
-// chatRoutes
-//   .route("/classroom/:id/messages")
-//   .post(protect, sendClassroomMessage)
-//   .get(protect, getClassroomMessages);
 
-chatRoutes.post("/classroom/join", protect, joinClassroom);
-chatRoutes.post("/classroom/leave", protect, leaveClassroom);
-
-chatRoutes.route("/classrooms").get(protect, getClassrooms);
 
 export default chatRoutes;
